Sequence and Collection Trait and their implementation in Deque, Stack and Vector
